# openapi-spec.yaml
openapi: 3.0.3
info:
  title: Student Assessment Platform API
  description: |
    REST API for the Student Assessment Platform - an educational web application for students (Class 1-10)
    to practice and excel in CBSE/ICSE board exams and competitive exams like Math Olympiad.

    ## Features
    - Practice and Mock Test modes
    - Gamification (badges, points, leaderboards)
    - Subscription management
    - Comprehensive admin panel
    - Configuration-driven content (add exams/subjects without code changes)

    ## Authentication
    All protected endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    - Authentication endpoints: 5 requests per 15 minutes
    - Student endpoints: 120 requests per minute
    - Admin endpoints: 300 requests per minute

    ## Pagination
    List endpoints support pagination using `page` and `limit` query parameters.
    Default: `page=1`, `limit=20`, Max: `limit=100`

    ## Error Handling
    All errors follow a standard format with `success: false`, `error` object containing `code` and `message`,
    and optional `details` array for validation errors.

  version: 1.0.0
  contact:
    name: API Support
    email: support@studentassessment.com
  license:
    name: Proprietary
servers:
  - url: https://api.studentassessment.com/api/v1
    description: Production server
...
    description: Staging server
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: JWT refresh token from login
                  example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
                      refreshToken:
                        type: string
                        description: New refresh token (rotated)
                        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
                      expiresIn:
                        type: integer
                        example: 900
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                error:
                  code: INVALID_REFRESH_TOKEN
                  message: Refresh token is invalid or expired
...
  /admin/configuration/subjects:
    get:
      tags:
        - Configuration
      summary: List all subjects (Admin)
      description: Get all subjects including inactive ones
      operationId: listSubjects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Subjects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      subjects:
                        type: array
                        items:
                          type: object
                          properties:
                            subjectId:
                              type: integer
                              example: 1
                            subjectName:
                              type: string
                              example: Mathematics
                            subjectCode:
                              type: string
                              example: MATH
                            description:
                              type: string
                              example: All mathematical topics
                            colorHex:
                              type: string
                              example: "#FF6B6B"
                            isActive:
                              type: boolean
                              example: true
                            displayOrder:
                              type: integer
                              example: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Configuration
      summary: Create new subject (Admin only)
      description: Add a new subject to the platform (configuration-driven)
      operationId: createSubject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subjectName
                - subjectCode
              properties:
                subjectName:
                  type: string
                  maxLength: 100
                  example: Science
                subjectCode:
                  type: string
                  maxLength: 50
                  example: SCIENCE
                description:
                  type: string
                  example: Physics, Chemistry, Biology
                colorHex:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
                  example: "#4CAF50"
                displayOrder:
                  type: integer
                  example: 2
      responses:
        '201':
          description: Subject created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Subject created successfully
                  data:
                    type: object
                    properties:
                      subjectId:
                        type: integer
                        example: 2
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Subject code already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login endpoint
...
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: Email is required
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid request data
              details:
                - field: email
                  message: Invalid email format
                - field: password
                  message: Password must be at least 8 characters
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: You don't have permission to perform this action
...